Refactor types to avoid circular dependency. (#1565)
As part of this change I remove the restriction that each test file must
correspond to a particular source file.
