+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Add API `isSimulcastSupported` so applications can check whether simulcast can be enabled and pass corresponding
-  policy.
+
+- Add API `isSimulcastSupported` so applications can check whether simulcast can be enabled and pass corresponding policy.
 - Add `bindToTileController` optional method to `VideoDownlinkBandwidthPolicy`.
   
 ### Removed
@@ -27,12 +27,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `VideoAdaptiveProbePolicy`.
 
 ### Changed
+
 - Allow passing in custom video simulcast uplink policy that implements the `SimulcastUplinkPolicy` interface.
 - Change the default video downlink policy to `VideoAdaptiveProbePolicy` to match with documentation.
 - Move configuration default from meeting session configuration to audio video controller.
-- Update the default priority-based video downlink policy to adjust target size based on number of videos in the 
-  meeting.
+- Update the default priority-based video downlink policy to adjust target size based on number of videos in the meeting.
 - Add a new section "Known Browser Issues" in FAQ.html.
+- Refactor some types to avoid a circular dependency (#1565).
   
 ## [2.18.0] - 2021-09-22
