+++ docs/classes/some.html
@@ -122,7 +122,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/maybeprovider.html">MaybeProvider</a>.<a href="../interfaces/maybeprovider.html#isnone">isNone</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/maybe/Some.ts#L9">src/maybe/Some.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/utils/Types.ts#L63">src/utils/Types.ts:63</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -133,7 +133,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/maybeprovider.html">MaybeProvider</a>.<a href="../interfaces/maybeprovider.html#issome">isSome</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/maybe/Some.ts#L8">src/maybe/Some.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/utils/Types.ts#L62">src/utils/Types.ts:62</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -151,7 +151,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/maybeprovider.html">MaybeProvider</a>.<a href="../interfaces/maybeprovider.html#defaulting">defaulting</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/maybe/Some.ts#L29">src/maybe/Some.ts:29</a></li>
+									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/utils/Types.ts#L83">src/utils/Types.ts:83</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -174,7 +174,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/maybe/Some.ts#L17">src/maybe/Some.ts:17</a></li>
+									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/utils/Types.ts#L71">src/utils/Types.ts:71</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-type-parameters-title">Type parameters</h4>
@@ -222,7 +222,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/maybeprovider.html">MaybeProvider</a>.<a href="../interfaces/maybeprovider.html#get">get</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/maybe/Some.ts#L21">src/maybe/Some.ts:21</a></li>
+									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/utils/Types.ts#L75">src/utils/Types.ts:75</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">T</span></h4>
@@ -240,7 +240,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/maybeprovider.html">MaybeProvider</a>.<a href="../interfaces/maybeprovider.html#getorelse">getOrElse</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/maybe/Some.ts#L25">src/maybe/Some.ts:25</a></li>
+									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/utils/Types.ts#L79">src/utils/Types.ts:79</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -263,7 +263,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/maybe/Some.ts#L13">src/maybe/Some.ts:13</a></li>
+									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/utils/Types.ts#L67">src/utils/Types.ts:67</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-type-parameters-title">Type parameters</h4>
@@ -310,7 +310,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/maybe/Some.ts#L33">src/maybe/Some.ts:33</a></li>
+									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/master/src/utils/Types.ts#L87">src/utils/Types.ts:87</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-type-parameters-title">Type parameters</h4>
