+++ src/contentsharecontroller/DefaultContentShareController.ts
@@ -6,12 +6,12 @@ import AudioVideoController from '../audiovideocontroller/AudioVideoController';
 import AudioVideoObserver from '../audiovideoobserver/AudioVideoObserver';
 import ContentShareObserver from '../contentshareobserver/ContentShareObserver';
 import Destroyable from '../destroyable/Destroyable';
-import Maybe from '../maybe/Maybe';
 import MeetingSessionConfiguration from '../meetingsession/MeetingSessionConfiguration';
 import MeetingSessionCredentials from '../meetingsession/MeetingSessionCredentials';
 import MeetingSessionStatus from '../meetingsession/MeetingSessionStatus';
 import DefaultModality from '../modality/DefaultModality';
 import AsyncScheduler from '../scheduler/AsyncScheduler';
+import { Maybe } from '../utils/Types';
 import VideoTile from '../videotile/VideoTile';
 import ContentShareConstants from './ContentShareConstants';
 import ContentShareController from './ContentShareController';
